updating README.md
Typo fixes (by invitation!) and other minor clarifications. 
I'm not 100% sure if "user stories" is the terminology we want to use here or if there is something else preferable, but adding some sort of subheading helps explain why the last three items seem to repeat some of the previous steps—and it allows the numbering to restart at 1 as intended.